backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-dotenv
langchain
langchain-groq
psycopg2-binary

